# openapi.yaml
openapi: 3.1.0
info:
  title: GitHub Issues Gateway
  version: 0.1.0
paths:
  /healthz:
    get:
      summary: Health Check
      operationId: health_check_healthz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /issues:
    get:
      summary: List Issues
      operationId: list_issues_issues_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      summary: Create Issue
      operationId: create_issue_issues_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /issues/{issue_number}:
    get:
      summary: Get Issue
      operationId: get_issue_issues__issue_number__get
      parameters:
      - name: issue_number
        in: path
        required: true
        schema:
          type: integer
          title: Issue Number
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update Issue
      operationId: update_issue_issues__issue_number__patch
      parameters:
      - name: issue_number
        in: path
        required: true
        schema:
          type: integer
          title: Issue Number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /issues/{issue_number}/comments:
    get:
      summary: List Comments
      operationId: list_comments_issues__issue_number__comments_get
      parameters:
      - name: issue_number
        in: path
        required: true
        schema:
          type: integer
          title: Issue Number
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Comment
      operationId: create_comment_issues__issue_number__comments_post
      parameters:
      - name: issue_number
        in: path
        required: true
        schema:
          type: integer
          title: Issue Number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /webhooks/github:
    post:
      summary: Github Webhook
      operationId: github_webhook_webhooks_github_post
      parameters:
      - name: x-hub-signature-256
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Hub-Signature-256
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /webhook-events:
    get:
      summary: List Webhook Events
      operationId: list_webhook_events_webhook_events_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    CommentCreate:
      properties:
        body:
          type: string
          title: Body
      type: object
      required:
      - body
      title: CommentCreate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IssueCreate:
      properties:
        title:
          type: string
          title: Title
        body:
          anyOf:
          - type: string
          - type: 'null'
          title: Body
      type: object
      required:
      - title
      title: IssueCreate
    IssueUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        body:
          anyOf:
          - type: string
          - type: 'null'
          title: Body
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
      type: object
      title: IssueUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError